+++ libs/nestjs/casl/src/guards/policies-guard.ts
@@ -11,7 +11,6 @@ import { CaslAbilityFactoryService } from '../services';
 import { isClass, PolicyHandlerType } from '@tractr/common';
 import {
   getRequestFromContext,
-  IS_PUBLIC_KEY,
   Logger,
   POLICIES_KEY,
 } from '@tractr/nestjs-core';
@@ -26,15 +25,6 @@ export class PoliciesGuard implements CanActivate {
   ) {}
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
-    // Check if the route is public
-    const isPublic = this.reflector.getAllAndOverride<boolean>(IS_PUBLIC_KEY, [
-      context.getHandler(),
-      context.getClass(),
-    ]);
-
-    // If we have the @Public decorator on our route we just let pass the request
-    if (isPublic) return true;
-
     const policyHandlers =
       this.reflector.get<PolicyHandlerType<unknown>[]>(
         POLICIES_KEY,