+++ libs/schematics/.jest/set-env-vars.js
@@ -1,0 +1,1 @@
+process.env.JEST_SKIP_INSTALL = true;