+++ libs/schematics/src/generators/init-workspace/generator.spec.ts
@@ -1,0 +1,16 @@
+import { readProjectConfiguration, Tree } from '@nrwl/devkit';
+import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
+
+import generator from './generator';
+
+describe('init-workspace generator', () => {
+  let appTree: Tree;
+
+  beforeEach(() => {
+    appTree = createTreeWithEmptyWorkspace();
+  });
+
+  it('should run successfully', async () => {
+    await generator(appTree);
+  });
+});