+++ libs/schematics/src/generators/init-workspace/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './add-files.helper';
+export * from './add-npmrc.helper';