+++ libs/schematics/src/generators/prettier-config/schema.d.ts
@@ -1,3 +1,4 @@
 export interface PrettierGeneratorSchema {
   format: boolean;
+  skipInstall?: boolean;
 }