+++ libs/schematics/src/generators/prisma-library/schema.d.ts
@@ -1,4 +1,5 @@
 export interface PrismaLibraryGeneratorSchema extends Record<string, unknown> {
   directory?: string;
   name: string;
+  skipInstall?: boolean;
 }