+++ libs/schematics/src/generators/release/schema.d.ts
@@ -1,3 +1,4 @@
 export interface ReleaseGeneratorSchema {
   project?: string;
+  skipInstall?: boolean;
 }